% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cofactorReport.R
\name{cofactorReport}
\alias{cofactorReport}
\title{cofactorReport}
\usage{
cofactorReport(intersectPeakMatrix, top_num = 10,
  cobinding_threshold = 0.05)
}
\arguments{
\item{intersectPeakMatrix}{Output of function 'intersectPeakMatrix()'.}

\item{top_num}{Number of highest co-binding factors to report for each TF (up to 10). By default the number is 10.}

\item{cobinding_threshold}{Only the co-factors with co-binding percentages more than this threshold value will be reported. By default the threshold is 0.05.}
}
\value{
A PDF file
}
\description{
This function allows you to get a PDF report of top cofactors along with DNA methylation for a TF.
}
\examples{
peak_id_x <- c("MM1_HSA_K562_CEBPB")
peak_id_y <- c("MM1_HSA_K562_CEBPD", "MM1_HSA_K562_ATF4")
intersect_output <- intersectPeakMatrix(peak_id_x=peak_id_x,
                                        motif_only_for_id_x=TRUE,
                                        peak_id_y=peak_id_y)
cofactorReport(intersectPeakMatrix = intersect_output)
}
\keyword{cofactorReport}
