% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonPeakResult.R
\name{commonPeakResult}
\alias{commonPeakResult}
\title{commonPeaks result}
\usage{
commonPeakResult(commonPeaks, return_common_peak_sites = FALSE,
  save_MethMotif_logo = FALSE, return_summary = FALSE,
  logo_type = "entropy", meth_level = "all")
}
\arguments{
\item{commonPeaks}{Required. commonPeaks() output, a matrix of CommonPeaksMM class objects.}

\item{return_common_peak_sites}{Optional. Either TRUE of FALSE (default). If TRUE, a list of data.frames containing common peak sets derived from "target_peak_list" will be returned. If "return_summary" is also TRUE, both will be returned in a list().}

\item{save_MethMotif_logo}{Optional. Either TRUE of FALSE (default). If TRUE, MethMotif logos for the common peak sets will be saved.}

\item{return_summary}{Optional. Either TRUE of FALSE (default). If TRUE, a numeric matrix containing the percentage of peaks in "target_peak_list" as common will be returned. If "return_common_peak_sites" is also TRUE, both will be returned in a list().}

\item{logo_type}{Optional. Logo type for the MethMotif logo to be saved, either "entropy" (default) or "frequency",}

\item{meth_level}{Optional. Methylation level to be plot for the MethMotif logo, and it should be one of the values, "all" (default), "methylated", and "unmethylated".}
}
\value{
a list of data.frames, a numeric matrix or MethMotif logo PDF files depending on the options.
}
\description{
This function allows you to get the reuslts from the commonPeaks() output, including a list of common peak sets, MethMotif logos, and common peak summary.
}
\examples{
target_peaks <- list(loadPeaks(id = "MM1_HSA_K562_CEBPB"),
                     read.delim("my_own_peaks.bed", header = F))
target_id <- c("MM1_HSA_K562_CEBPB", "my_own_peaks")
compared_peaks <- list(loadPeaks(id = "MM1_HSA_HepG2_CEBPB"),
                      read.delim("peaks_to_common_with.bed", header = F))
compared_id <- c("MM1_HSA_HepG2_CEBPB", "peaks_to_common_with")
commonPeaks_output <- commonPeaks(target_peak_list=target_peaks,
                                  target_peak_id=target_id,
                                  compared_peak_list=compared_peaks,
                                  compared_peak_id=compared_id)
commonPeaks_result <- commonPeakResult(commonPeaks=commonPeaks_output,
                                       return_exclusive_peak_sites=T,
                                       save_MethMotif_logo=T,
                                       return_summary=T)
}
\keyword{commonPeakResult}
