% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBrowser.R
\name{dataBrowser}
\alias{dataBrowser}
\title{Browse the current data available in TFregulomeR}
\usage{
dataBrowser(species, organ, sample_type, cell_tissue_name, tf,
  disease_state, source, server = "sg", TFregulome_url)
}
\arguments{
\item{species}{The species of interset}

\item{organ}{The organ of interset}

\item{sample_type}{The sample type of interset}

\item{cell_tissue_name}{The name of tissue or cell of interset}

\item{tf}{The TF of interset}

\item{disease_state}{The disease state of interset}

\item{source}{The source of interset}

\item{server}{server localtion to be linked, either 'sg' or 'ca'.}

\item{TFregulome_url}{TFregulomeR server is implemented in MethMotif server. If the MethMotif url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/" or "http://methmotif.org", please use a new url.}
}
\value{
data.frame containing the information of the queried TFBSs in TFregulomeR
}
\description{
This function allows you to get the current data in TFregulomeR
}
\examples{
TFBS_brain <- dataBrowser(organ = "brain")
}
\keyword{ChIPseq}
\keyword{data}
