% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusivePeaks.R
\name{exclusivePeaks}
\alias{exclusivePeaks}
\title{exclusivePeaks}
\usage{
exclusivePeaks(target_peak_list, target_peak_id, excluded_peak_list,
  excluded_peak_id, motif_type = "MEME", methmotif_url)
}
\arguments{
\item{target_peak_list}{Required. List of data.frames, each of which contains bed-format peak regions. They are the peak sets you want to get the exclusive subsets from, and can be loaded from MethMotif database or self-provided.}

\item{target_peak_id}{Required. Character of vector, each of which is a unique ID corresponding to the element in "target_peak_list". If a peak set is from MethMotif Database, its MethMotif ID should be used here.}

\item{excluded_peak_list}{Required. List of data.frames, each of which contains bed-format peak regions. They are the peak sets you want to exclude from the "target_peak_list", and can be loaded from MethMotif database or self-provided.}

\item{excluded_peak_id}{Required. Character of vector, each of which is a unique ID corresponding to the element in "excluded_peak_list". If a peak set is from MethMotif Database, its MethMotif ID should be used here.}

\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}

\item{motif_format}{Required. Motif PFM format, either in MEME by default or TRANSFAC.}
}
\value{
matrix of ExclusivePeaksMM class objects
}
\description{
This function allows you to obtain the subsets of a list of peak sets unique to another list of peak sets (either from MethMotif database or the self-provided).
}
\examples{
target_peaks <- list(loadPeaks(id = "MM1_HSA_K562_CEBPB"),
                     read.delim("my_own_peaks.bed", header = F))
target_id <- c("MM1_HSA_K562_CEBPB", "my_own_peaks")
excluded_peaks <- list(loadPeaks(id = "MM1_HSA_HepG2_CEBPB"),
                       read.delim("peaks_to_exclude.bed", header = F))
excluded_id <- c("MM1_HSA_HepG2_CEBPB", "peaks_to_exclude")
exclusivePeaks_output <- exclusivePeaks(target_peak_list=target_peaks,
                                        target_peak_id=target_id,
                                        excluded_peak_list=excluded_peaks,
                                        excluded_peak_id=excluded_id)
}
\keyword{exclusivePeaks}
