% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeAnnotate.R
\name{genomeAnnotate}
\alias{genomeAnnotate}
\title{genomeAnnotate}
\usage{
genomeAnnotate(peaks, assembly = "hg38", return_annotation = FALSE,
  return_html_report = FALSE, promoter_range = c(-1000, 100),
  TTS_range = c(-100, 1000), TFregulome_url)
}
\arguments{
\item{peaks}{Required. A bed-format genomic regions in data frame.}

\item{assembly}{The genome assembly of the input regions, currently supporting 'hg19', 'hg38' (default), 'mm9' and 'mm10'.}

\item{return_annotation}{Either TRUE of FALSE (default). If TRUE, a data.frame containing annotation results will be returned.}

\item{return_html_report}{Either TRUE of FALSE (default). If TRUE, a dynamic HTML report will be saved.}

\item{promoter_range}{A numeric vector to define promoter range. By default, c(-1000, 100) defines promoters as 1000bp upstream and 100bp downstream of TSS.}

\item{TTS_range}{A numeric vector to define TTS range. By default, c(-100, 1000) defines promoters as 100bp upstream and 1000bp downstream of real TTS.}

\item{TFregulome_url}{TFregulome server is implemented in MethMotif server. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
a data.frame, or an HTML report depending on the options.
}
\description{
This function allows you to annotate genomic locations of cis-regulatory regions.
}
\examples{
require(TxDb.Hsapiens.UCSC.hg38.knownGene)
HCT116_CEBPB_regions <- loadPeaks(id = "MM1_HSA_HCT116_CEBPB", includeMotifOnly=TRUE)
HCT116_CEBPB_regions_annotation <- genomeAnnotate(peaks = HCT116_CEBPB_regions[1:100,],
                              return_annotation = TRUE, return_html_report = TRUE)
}
\keyword{genomeAnnotate}
