% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greatAnnotate.R
\name{greatAnnotate}
\alias{greatAnnotate}
\title{greatAnnotate}
\usage{
greatAnnotate(peaks, assembly = "hg38", return_annotation = FALSE,
  return_html_report = FALSE, pvalue = 0.01, test = "binomial",
  great_rule = "basalPlusExt", great_adv_upstream = 5,
  great_adv_downstream = 1, great_adv_span = 1000,
  great_adv_twoDistance = 1000, great_adv_oneDistance = 1000,
  request_interval = 60, great_version = 3)
}
\arguments{
\item{peaks}{Required. A bed-format genomic regions in data frame.}

\item{assembly}{The genome assembly of the input regions, currently supporting 'hg19', 'hg38' (default), 'mm9' and 'mm10'.}

\item{return_annotation}{Either TRUE of FALSE (default). If TRUE, a data.frame containing annotation results will be returned.}

\item{return_html_report}{Either TRUE of FALSE (default). If TRUE, a dynamic HTML report will be saved.}

\item{pvalue}{The adjusted p-value which is applied to filter the results, by default 0.01.}

\item{test}{The statistical test used in GREAT analysis, either 'binomial' (default) or 'hypergeometric'.}

\item{great_rule}{Equivalent to the rGREAT input 'rule', 'basalPlusExt' (default, basal plus extension), 'twoClosest' (two nearest genes), or 'oneClosest' (single nearest gene).}

\item{great_adv_upstream}{Equivalent to the rGREAT input 'adv_upstream' (upstream extension, kb). Only applicable when 'great_rule' is  'basalPlusExt', by default 5.0 (kb).}

\item{great_adv_downstream}{Equivalent to the rGREAT input 'adv_downstream' (downstream extension, kb). Only applicable when 'great_rule' is  'basalPlusExt', by default 1.0 (kb).}

\item{great_adv_span}{Equivalent to the rGREAT input 'adv_span' (maximal distal region). Only applicable when 'great_rule' is  'basalPlusExt', by default 1000.0 (kb).}

\item{great_adv_twoDistance}{Equivalent to the rGREAT input 'adv_twoDistance' (region range to be considered). Only applicable when 'great_rule' is  'twoClosest', by default 1000.0 (kb).}

\item{great_adv_oneDistance}{Equivalent to the rGREAT input 'adv_oneDistance' (region range to be considered). Only applicable when 'great_rule' is  'oneClosest', by default 1000.0 (kb).}

\item{request_interval}{The minimal gap time between two requests using greatAnnotate, by default 60 (s).}

\item{great_version}{Equivalent to the rGREAT input 'version', by default 3.0.}
}
\value{
a data.frame, or an HTML report depending on the options.
}
\description{
This function allows you to analyse the gene ontologies of targeting genes by cis-regulatory regions.
}
\examples{
require(rGREAT)
require(liftOver)
require(rbokeh)
K562_CEBPB_regions <- loadPeaks(id = "MM1_HSA_K562_CEBPB")
K562_CEBPB_regions_annotation <- greatAnnotate(peaks = K562_CEBPB_regions[1:500,],
                             return_annotation = TRUE, return_html_report = TRUE)
}
\keyword{greatAnnotate}
