% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactome3D.R
\name{interactome3D}
\alias{interactome3D}
\title{interactome3D}
\usage{
interactome3D(intersectPeakMatrix, return_interactome_with_mCpG = FALSE,
  mCpG_threshold = 0.8,
  return_interactome_with_external_source = FALSE,
  external_source_value = "median", angle_of_matrix = "x")
}
\arguments{
\item{intersectPeakMatrix}{Output of function 'intersectPeakMatrix()'.}

\item{return_interactome_with_mCpG}{Either TRUE of FALSE (default). If TRUE, html report of TF interactome with mCpG portion will be saved.}

\item{mCpG_threshold}{A mininum beta score to determine CpG methylation. Should be 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8 (default), or 0.9}

\item{return_interactome_with_external_source}{Either TRUE of FALSE (default). If TRUE, html report of TF interactome with external source signal will be saved.}

\item{external_source_value}{The value of external source signal in the intersected peaks. It should be one of the following values: "median" (default),"mean","SD","quartile_25","quartile_75".}

\item{angle_of_matrix}{Either "x" (default) or "y". If "x", will focus on the peak sets in "peak_list_x" intersected with "peak_list_y"; if "y", will focus on peak sets in "peak_list_y" intersected with "peak_list_x".}
}
\value{
An html file
}
\description{
This function allows you to get a html report of a 3D dynamic TF interactome with CpG methylation and external source signal.
}
\keyword{interactome3D}
