% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersectPeakMatrix.R
\name{intersectPeakMatrix}
\alias{intersectPeakMatrix}
\title{intersectPeakMatrix}
\usage{
intersectPeakMatrix(peak_list_x, peak_list_y, peak_id_x, peak_id_y,
  motif_type = "MEME", methmotif_url)
}
\arguments{
\item{peak_list_x}{Required. List of data.frames, each of which contains bed-format peak regions. They are the peak sets you want to intersect with each of "peak_list_y", and can be loaded from MethMotif database or self-provided.}

\item{peak_list_y}{Required. List of data.frames, each of which contains bed-format peak regions. They are the peak sets you want to intersect with each of "peak_list_x", and can be loaded from MethMotif database or self-provided.}

\item{peak_id_x}{Required. Character of vector, each of which is a unique ID corresponding to the element in "peak_list_x". If a peak set is from MethMotif Database, its MethMotif ID should be used here.}

\item{peak_id_y}{Required. Character of vector, each of which is a unique ID corresponding to the element in "peak_list_y". If a peak set is from MethMotif Database, its MethMotif ID should be used here.}

\item{motif_type}{Required. Motif PFM format, either in MEME by default or TRANSFAC.}

\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
matrix of IntersectPeakMatrix class objects
}
\description{
This function allows you to obtain a numeric matrix denoting the pair-wise intersecting percentages between two lists of peak sets, as well as MethMotif logos for each pair of intersections if the original peak set is loaded from MethMotif database.
}
\examples{
peak_list_x <- list(loadPeaks(id = "MM1_HSA_K562_CEBPB"),
                    read.delim("my_own_peak1.bed", header = F))
peak_id_x <- c("MM1_HSA_K562_CEBPB", "my_own_peak1")
peak_list_y <- list(loadPeaks(id = "MM1_HSA_HepG2_CEBPB"),
                    read.delim("my_own_peak2.bed", header = F))
peak_id_y <- c("MM1_HSA_HepG2_CEBPB", "my_own_peak2")
intersectPeakMatrix_output <- intersectPeakMatrix(peak_list_x=peak_list_x,
                                                  peak_list_y=peak_list_y,
                                                  peak_id_x=peak_id_x,
                                                  peak_id_y=peak_id_y)
}
\keyword{intersectPeakMatrix}
