% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCell.R
\name{listCell}
\alias{listCell}
\title{the existing cell types in MethMotif database for a given species}
\usage{
listCell(species = "human", methmotif_url)
}
\arguments{
\item{species}{Required. The species of interset, default human}

\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
character vector giving the existing cell types for a given species in MethMotif database
}
\description{
This function allows you to get existing cell types for a given species in MethMotif database.
}
\examples{
listCell()
listCell(species = "mouse")
}
\keyword{cell}
