% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSpecies.R
\name{listSpecies}
\alias{listSpecies}
\title{the existing species in MethMotif database}
\usage{
listSpecies(methmotif_url)
}
\arguments{
\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
character vector giving the existing species in MethMotif database
}
\description{
This function allows you to get existing species in MethMotif database.
}
\examples{
listSpecies()
}
\keyword{species}
