% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTF.R
\name{listTF}
\alias{listTF}
\title{the existing TFBSs in MethMotif database}
\usage{
listTF(cell_type, methmotif_url)
}
\arguments{
\item{cell_type}{Required. The cell types of interset}

\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
data.frame giving the existing TFBSs in MethMotif database for a given cell type
}
\description{
This function allows you to get existing TFBSs for a given cell type in MethMotif database.
}
\examples{
listTF(cell_type = "K562")
}
\keyword{TFBS}
