% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifDistrib.R
\name{motifDistrib}
\alias{motifDistrib}
\title{profile TFBS distribution}
\usage{
motifDistrib(id, peak_list, peak_id, plot_of_each_side = 100,
  methmotif_url)
}
\arguments{
\item{id}{Required. MethMotif ID}

\item{peak_list}{Required. List of data.frames, each of which contains bed-format peak regions. They are the peak sets you want to profile TFBS distributions in, and can be loaded from MethMotif database or self-provided.}

\item{peak_id}{Required. Character of vector, each of which is a unique ID corresponding to the element in "peak_list". If a peak set is directly loaded from MethMotif Database, its MethMotif ID should be used here.}

\item{plot_of_each_side}{Optional. By default 100bp, and motif occurrences in a window of +/- 100bp around peak centres will be returned.}

\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
a list containing the numbers of input peaks and peaks with motif, as well as motif occurrences in the plot window.
}
\description{
This function allows you to profile TFBS distributions in a given list of peak sets.
}
\examples{
CEBPB_peaks <- loadPeaks(id = "MM1_HSA_K562_CEBPB")
motifDistrib_output <- motifDistrib(id = "MM1_HSA_K562_CEBPB",
                                    peak_list = list(CEBPB_peaks),
peak_id = c("MM1_HSA_K562_CEBPB"))
}
\keyword{motifDistrib}
