% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistrib.R
\name{plotDistrib}
\alias{plotDistrib}
\title{plot TFBS distribution from motifDistrib() output}
\usage{
plotDistrib(motifDistrib)
}
\arguments{
\item{motifDistrib}{Required. motifDistrib() output.}
}
\value{
TFBS distribution PDF file.
}
\description{
This function allows you to plot TFBS distributions in a given list of peak sets from the output of motifDistrib().
}
\examples{
CEBPB_peaks <- loadPeaks(id = "MM1_HSA_K562_CEBPB")
motifDistrib_output <- motifDistrib(id = "MM1_HSA_K562_CEBPB",
                                    peak_list = list(CEBPB_peaks[1:100,]),
                                    peak_id = c("MM1_HSA_K562_CEBPB"))
plotDistrib(motifDistrib = motifDistrib_output)
}
\keyword{plotDistrib}
