% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLogo.R
\name{plotLogo}
\alias{plotLogo}
\title{plot (Meth)Motif logo}
\usage{
plotLogo(MM_object, logo_type = "entropy", meth_level = "all")
}
\arguments{
\item{MM_object}{Required. MethMotif class object}

\item{logo_type}{Logo type for the (Meth)Motif logo to be saved, either "entropy" (default) or "frequency".}

\item{meth_level}{Methylation level to be plot for the (Meth)Motif logo, and it should be one of the values, "all" (default), "methylated", and "unmethylated".}
}
\value{
(Meth)Motif logo pdf file. If the TFregulome peak source is from MethMotif, MethMotif logo will be saved; if the source is GTRD, only motif logo will be saved.
}
\description{
This function allows you to plot (Meth)Motif logo.
}
\examples{
K562_CEBPB <- searchMotif(id = "MM1_HSA_K562_CEBPB")
plotLogo(MM_object = K562_CEBPB)
}
\keyword{plotLogo}
