% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMethMotif.R
\name{searchMethMotif}
\alias{searchMethMotif}
\title{Motif PFM and beta score matrix for a given MethMotif ID or transcription factor}
\usage{
searchMethMotif(tf, id, cell, motif_format = "MEME", methmotif_url)
}
\arguments{
\item{tf}{Required either this or "id". Transcription factor name, case insensitive.}

\item{id}{Required either this or "tf". MethMotif ID, case insensitive. If "id" is provided, other inputs, namely "tf" and/or "cell", will be ignored.}

\item{cell}{Optional. Cell type of interest which will be combined with tf, case insensitive.}

\item{motif_format}{Required. Motif PFM format, either in MEME by default or TRANSFAC.}

\item{methmotif_url}{Optional. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
vector of MethMotif class objects with each element named with a MethMotif ID
}
\description{
This function allows you to obtain motif PFM matrix and beta score matrix for a given MethMotif ID or transcription factor.
}
\examples{
K562_CEBPB <- searchMethMotif(id = "MM1_HSA_K562_CEBPB")
K562_CEBPB <- searchMethMotif(tf = "cebpb", cell = "k562")
K562_CEBPB_transfac <- searchMethMotif(id = "MM1_HSA_K562_CEBPB",
                                       motif_format = "TRANSFAC")
CEBPB <- searchMethMotif(tf = "MM1_HSA_K562_CEBPB")
}
\keyword{MethMotif}
