% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMotif.R
\name{searchMotif}
\alias{searchMotif}
\title{Search motif PFM and beta score matrix (if source is MethMotif) for a given TFregulomeR ID in TFregulomeR}
\usage{
searchMotif(id, motif_format = "MEME", server = "sg", TFregulome_url)
}
\arguments{
\item{id}{Required. TFregulomeR ID.}

\item{motif_format}{Motif PFM format, either in MEME by default or TRANSFAC.}

\item{server}{server localtion to be linked, either 'sg' or 'ca'.}

\item{TFregulome_url}{TFregulomeR server is implemented in MethMotif server. If the MethMotif url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/" or "http://methmotif.org", please use a new url.}
}
\value{
MethMotif class object
}
\description{
This function allows you to obtain motif PFM matrix and beta score matrix (if source is MethMotif) for a given TFregulomeR ID in TFregulomeR
}
\examples{
K562_CEBPB <- searchMotif(id = "MM1_HSA_K562_CEBPB")
K562_CEBPB_transfac <- searchMotif(id = "MM1_HSA_K562_CEBPB",
                                   motif_format = "TRANSFAC")
}
\keyword{MethMotif}
