% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonPeakResult.R
\name{commonPeakResult}
\alias{commonPeakResult}
\title{commonPeaks result}
\usage{
commonPeakResult(commonPeaks, return_common_peak_sites = FALSE,
  save_MethMotif_logo = FALSE, return_methylation_profile = FALSE,
  return_summary = FALSE, logo_type = "entropy", meth_level = "all")
}
\arguments{
\item{commonPeaks}{Required. commonPeaks() output, a matrix of CommonPeaksMM class objects.}

\item{return_common_peak_sites}{Either TRUE of FALSE (default). If TRUE, a list of data.frames containing common peak sets.}

\item{save_MethMotif_logo}{Either TRUE of FALSE (default). If TRUE, (Meth)Motif logos for the common peak sets will be saved.}

\item{return_methylation_profile}{Either TRUE of FALSE (default). If TRUE, the methylation profiles in 200bp window around common peak summits will be returned.}

\item{return_summary}{Either TRUE of FALSE (default). If TRUE, a numeric matrix containing the percentage of peaks as common will be returned.}

\item{logo_type}{Logo type for the (Meth)Motif logo to be saved, either "entropy" (default) or "frequency",}

\item{meth_level}{Methylation level to be plotted for the (Meth)Motif logo, and it should be one of the values, "all" (default), "methylated", and "unmethylated".}
}
\value{
a list of data.frames, a numeric matrix or (Meth)Motif logo PDF files depending on the options.
}
\description{
This function allows you to get the reuslts from the commonPeaks() output,
including a list of common peak sets, (Meth)Motif logos, methylation profile
in common peaks and common peak summary.
}
\examples{
target_id <- c("MM1_HSA_K562_CEBPB")
compared_id <- c("MM1_HSA_HepG2_CEBPB")
commonPeaks_output <- commonPeaks(target_peak_id=target_id,
                                  motif_only_for_target_peak=TRUE,
                                  compared_peak_id=compared_id,
                                  motif_only_for_compared_peak=TRUE,
                                  methylation_profile_in_narrow_region=TRUE)
commonPeaks_result <- commonPeakResult(commonPeaks=commonPeaks_output,
                                       return_common_peak_sites=TRUE,
                                       save_MethMotif_logo=TRUE,
                                       return_methylation_profile=TRUE,
                                       return_summary=TRUE)
}
\keyword{commonPeakResult}
