% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusivePeaks.R
\name{exclusivePeaks}
\alias{exclusivePeaks}
\title{exclusivePeaks}
\usage{
exclusivePeaks(target_peak_id, motif_only_for_target_peak = FALSE,
  user_target_peak_list, user_target_peak_id, excluded_peak_id,
  motif_only_for_excluded_peak = FALSE, user_excluded_peak_list,
  user_excluded_peak_id, methylation_profile_in_narrow_region = TRUE,
  motif_type = "MEME", server = "sg", TFregulome_url)
}
\arguments{
\item{target_peak_id}{Character of vector, each of which is a TFregulomeR ID. Each of target peak will be compared with all "excluded peaks" to get its exclusive subset.}

\item{motif_only_for_target_peak}{Either TRUE of FALSE (default). If TRUE, only peaks with motif will be loaded for each TFregulomeR ID in target_peak_id.}

\item{user_target_peak_list}{A list of data.frames, each of which contains user's own bed-format target peak regions.}

\item{user_target_peak_id}{Character of vector, each of which is a unique ID corresponding to each peak set in the list user_target_peak_list. If the IDs are not provided or not unique, the function will automatically generate the IDs of its own. If any of the peak sets is derived from TFregulomeR database, its TFregulomeR ID should be used here correspondingly.}

\item{excluded_peak_id}{Character of vector, each of which is a TFregulomeR ID.}

\item{motif_only_for_excluded_peak}{Either TRUE of FALSE (default). If TRUE, only peaks with motif will be loaded for each TFregulomeR ID in excluded_peak_id.}

\item{user_excluded_peak_list}{A list of data.frames, each of which contains user's own bed-format excluded peak regions.}

\item{user_excluded_peak_id}{Character of vector, each of which is a unique ID corresponding to each peak set in the list user_excluded_peak_list. If the IDs are not provided or not unique, the function will automatically generate the IDs of its own. If any of the peak sets is derived from TFregulomeR database, its TFregulomeR ID should be used here correspondingly.}

\item{methylation_profile_in_narrow_region}{Either TRUE (default) of FALSE. If TRUE, methylation states in 200bp window surrounding peak summits for each exclusive peak from target_peak_id and user_target_peak_list (with TFregulomeR ID).}

\item{motif_type}{Motif PFM format, either in MEME by default or TRANSFAC.}

\item{server}{server localtion to be linked, either 'sg' or 'ca'.}

\item{TFregulome_url}{TFregulomeR server is implemented in MethMotif server. If the MethMotif url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/" or "http://methmotif.org", please use a new url.}
}
\value{
matrix of ExclusivePeaksMM class objects
}
\description{
This function allows you to obtain a list of exclusive peak subsets along with the DNA methylation profiles.
}
\examples{
target_id <- "MM1_HSA_K562_CEBPB"
excluded_id <- c("MM1_HSA_HepG2_CEBPB", "MM1_HSA_HCT116_CEBPB")
excluPeak_output <- exclusivePeaks(target_peak_id=target_id,
                                   motif_only_for_target_peak=TRUE,
                                   excluded_peak_id=excluded_id,
                                   motif_only_for_excluded_peak=TRUE,
                                   methylation_profile_in_narrow_region=TRUE)
}
\keyword{exclusivePeaks}
