% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFBSBrowser.R
\name{TFBSBrowser}
\alias{TFBSBrowser}
\title{Browse the current TFBSs available in TFregulome database}
\usage{
TFBSBrowser(species, organ, sample_type, cell_tissue_name, tf,
  disease_state, source, TFregulome_url)
}
\arguments{
\item{species}{The species of interset}

\item{organ}{The organ of interset}

\item{sample_type}{The sample type of interset}

\item{cell_tissue_name}{The name of tissue or cell of interset}

\item{tf}{The TF of interset}

\item{disease_state}{The disease state of interset}

\item{source}{The source of interset}

\item{TFregulome_url}{TFregulome server is implemented in MethMotif server. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
data.frame containing the information of the queried TFBSs in TFregulome database
}
\description{
This function allows you to get the current TFBSs in TFregulome database
}
\examples{
TFBS_brain <- TFBSBrowser(organ = "brain")
}
\keyword{TFBS}
