% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toTFBSTools.R
\name{toTFBSTools}
\alias{toTFBSTools}
\title{convert motif PFM in TFregulome database into PFMatrix class object in TFBSTools package}
\usage{
toTFBSTools(id, TFregulome_url)
}
\arguments{
\item{id}{Required. TFregulome ID.}

\item{TFregulome_url}{TFregulome server is implemented in MethMotif server. If the MethMoitf url is NO more "http://bioinfo-csi.nus.edu.sg/methmotif/", please use a new url.}
}
\value{
An object of class PFMatrix
}
\description{
This function allows you to retrieve and convert motif PFM in TFregulome database into PFMatrix class object, which can be used in TFBSTools package.
}
\examples{
require(TFBSTools)
CEBPB_pfm <- toTFBSTools(id = "MM1_HSA_K562_CEBPB")
}
\keyword{toTFBSTools}
